test(bit_iterator): compatibility test with bitstream crate

+++ src/bit_iterator.rs
@@ -28,7 +28,7 @@ where
 
     fn next(&mut self) -> Option<Self::Item> {
         loop {
-            let bit = self.i % self.n;
+            let bit= (self.i % self.n) as u8;
             self.i += 1;
             if bit == 0 {
                 self.byte = None;
